Warnings after done_testing() should not cause cause test failures
To demonstrate the problem, I simply plumb in a call to `warn` after
done_testing(), which causes a test failure for the existing code;
while this is consistent with the intent of the plugin (if maybe a
little above-and-beyond), it's a test failure because of a bad plan,
which suggests that it's not the right way to handle the situation.

[The `Test2::API` documentation|https://metacpan.org/pod/Test2::API#$bool-=-test2_is_testing_done()]
happens to address exactly this scenario, so we borrow their solution.

Since all other behavior the `$SIG{__WARN__}` handler displays
continues to be appropriate, we simply avoid the `context_do` call if
testing has completed.

This resolves issue #2.